requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==21.2.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic[email]==2.9.2
passlib==1.7.4
bcrypt==4.0.1
pyjwt==2.9.0
python-multipart==0.0.12
reportlab==4.2.5
aiofiles==23.2.1
